@@ -7,9 +7,10 @@
     "gh-pages": "^3.1.0",
     "materialize": "^1.0.0",
     "materialize-css": "^1.0.0-rc.2",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-dotenv": "^0.1.3",
+    "react-id-generator": "^3.0.2",
     "react-router": "^6.16.0",
     "react-router-dom": "^6.16.0",
     "react-scripts": "^5.0.1"
