@@ -1,21 +1,21 @@
-import React, {useEffect, useState} from 'react'
-import firebase from './firebase.js'
+import React, {createContext, useEffect, useState} from 'react'
+import {auth, onAuthStateChanged} from './firebase.js'
 
-export const AuthContext = React.createContext()
+export const AuthContext = createContext()
 export const AuthProvider = ({children}) => {
     const [currUser, setCurrUser] = useState(null)
     const [pending, setPending] = useState(true)
     useEffect(() => {
-        firebase.auth().onAuthStateChanged((user) => {
+        const unsubscribe = onAuthStateChanged(auth, (user) => {
             setCurrUser(user)
             setPending(false)
-        })
-
+          })
+          return () => unsubscribe()
     }, [])
-    if (pending) {
+    if(pending) {
       return (
-        <div id="centerScreen" className="preloader-wrapper active">
-          <div className="spinner-layer spinner-red-only">
+        <div className="preloader-wrapper active | flex mx-auto mt-[15%]">
+          <div className="spinner-layer spinner-green-only">
             <div className="circle-clipper left">
               <div className="circle"></div>
             </div>
@@ -29,7 +29,6 @@ export const AuthProvider = ({children}) => {
         </div>
       );
     }
-
     return (
         <AuthContext.Provider value={{currUser}}>
             {children}
