@@ -4,21 +4,22 @@ const cors = require('cors');
 const twilio = require("twilio");
 const client = twilio(
   process.env.TWILIO_ACCOUNT_SID,
-  process.env.TWILIO_AUTH_KEY 
+  process.env.TWILIO_AUTH_KEY
 );
 const nodemailer = require("nodemailer")
 const express = require("express");
 const app = express();
-app.use(express.urlencoded({ extended: true })); 
+app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 app.use(cors())
 
+
 const firebase = require("firebase/app");
-require("firebase/firestore");
+const { collection, deleteDoc, query, where, getDocs } = require("firebase/firestore");
 
 const config = {
   apiKey: process.env.FIREBASE_API_KEY,
-  authDomain: process.env.AUTH_DOMAIN, 
+  authDomain: process.env.AUTH_DOMAIN,
   databaseURL: process.env.DB_URL,
   projectId: "reminder-app-4be61",
   storageBucket: "reminder-app-4be61.appspot.com",
@@ -27,37 +28,33 @@ const config = {
   measurementId: "G-T600ZD0201",
 };
 firebase.initializeApp(config);
+function convertToUTC(inputDate, inputTime, inputTimezone) {
+  // Define a mapping of timezone options to their respective offsets in minutes
+  const timezoneOffsets = {
+    "US/Eastern": -300, // UTC-5
+    "US/Central": -360, // UTC-6
+    "US/Mountain": -420, // UTC-7
+    "US/Pacific": -480, // UTC-8
+    "US/Hawaii": -600, // UTC-10
+    "US/Alaska": -540, // UTC-9
+  };
 
-const timeFactory = (hour, midday) => {
-  if (midday === "AM" && hour === "12") {
-    hour = "00";
-  } else if (midday === "PM" && hour === "12") {
-    hour = "12";
-  } else if (midday === "PM" && Number(hour) < 12) {
-    hour = Number(hour) + 12;
-  } else if (midday === "AM" && Number(hour) < 12) {
-    if (hour[0] === "0") {
-      hour = 0 + Number(hour);
-      hour = "0" + String(hour);
-    } else {
-      hour = 0 + Number(hour);
-    }
-  }
-  return String(hour);
-};
-year = ""
-month = ""
-day = ""
-const dateFactory = (date) => {
-    console.log('date -->',date)
-    const dateObj = new Date(date);
-    day = dateObj.getDate();
-    console.log('day -->',day)
-    month = dateObj.getMonth()+1;
-    console.log('month -->',month)
-    year = dateObj.getFullYear();
-    console.log('year -->',year)
-};
+  // Create a new Date object using the provided date and time
+  const dateTimeString = `${inputDate} ${inputTime}`;
+  const localDate = new Date(dateTimeString);
+
+  // Get the offset for the specified timezone
+  const timezoneOffset = timezoneOffsets[inputTimezone];
+
+  // Calculate the UTC time in milliseconds since January 1, 1970
+  const utcTime = localDate.getTime() - (localDate.getTimezoneOffset() + timezoneOffset) * 60000;
+
+  // Create a new Date object with the UTC time
+  const utcDate = new Date(utcTime);
+
+  return utcDate;
+}
+
 app.post('/createMessage', (req, res) => {
   // Call the desired function here
   try {
@@ -68,43 +65,56 @@ app.post('/createMessage', (req, res) => {
       time,
       date,
       message,
-      //timezone *Should be used in conversion
+      timezone,
+      ID
     } = req.body
     const fetchData = async () => {
-      let timeframe = time.slice(6, 8);
-      let minute = time.slice(3, 5);
-      let hour = time.slice(0, 2);
-      hour = timeFactory(hour, timeframe);
-      dateFactory(date)
-      //UTC conversion func
-      const localDate = new Date(year, month - 1, day, hour, minute);
-      const utcDate = new Date(new Date().getTime() + 24 * 60 * 60 * 1000)//new Date(localDate.toLocaleString("en-US", { timeZone: "UTC" }));
+      const inputDate = date;
+      const inputTime = time;
+      const inputTimezone = timezone; // Change this to your desired timezone
 
-      console.log("local -->",localDate, "utc -->", utcDate, "after string func -->", utcDate.toISOString());
-      console.log('ph -->',phoneNumber)
+      const utcDate = convertToUTC(inputDate, inputTime, inputTimezone);
+
+      console.log("time -->", utcDate.toISOString());
+      console.log('ph -->', phoneNumber)
+
       await client.messages.create(
         {
-          from: '+18886648381',
+          from: process.env.TWILIO_NUMBER,
           messagingServiceSid: process.env.TWILIO_MESSAGING_SERVICE_SID,
-          body: `Hello ${name},\n ${message}`,
+          body: `${message}`,
           sendAt: utcDate.toISOString(),
           scheduleType: 'fixed',
           to: phoneNumber,
-        }
-        ,
-        (err, message) => {
-          console.log("BE MSG --> ",message.body);
-          console.log("Error -> ",err);
+        },
+        async (err, message) => {
+          if (err) {
+            console.log("Error -> ", err);
+            return;
+          }
+          console.log("BE MSG --> ", message.body);
+          console.log("id --> ", ID);
 
+          // const q = query(collection("reminders"), where("ID", "==", ID));
+
+          // try {
+          //   const querySnapshot = await getDocs(q);
+          //   querySnapshot.forEach(async (doc) => {
+          //     // Delete the document using the deleteDoc function
+          //     await deleteDoc(doc.ref);
+          //     console.log('Document deleted successfully');
+          //   });
+          // } catch (error) {
+          //   console.error("Error deleting document:", error);
+          // }
         }
-      ).then(message => console.log(message))
-      .done();
+      ).done();
     }
     fetchData();
     // Send a response back to the frontend
     console.log(res.json({ message: 'BE --> Function triggered successfully' }))
   } catch (err) {
-    console.error("Error: ",err)
+    console.error("Error: ", err)
   }
 });
 
